Delete the old document after an update.
This commit makes sure that the old document is deleted
after an update. So that we do not duplicate documents.
Xapian supports a replace_document API but it seems to
create a new document anyway, so this commit switch back
to Xapian add_document delete_document API.

Introduce unit tests to check the logic of the db update.

Signed-off-by: Clement Verna <cverna@tutanota.com>